extend conforming classes
